Refactor header component and remove unused code
