--- frontend/app/common/header/index.ts
@@ -1,4 +1,0 @@
-export * from './Header'
-export * from './components/MobileCategories'
-export * from './components/MobileMenu'
-export * from './components/MobileSearch'
